Refactor to int :)

+++ Pong-2014/src/main/java/pong2014/logiikka/Kentta.java
@@ -7,33 +7,33 @@ package pong2014.logiikka;
 public class Kentta {
     
     private Pallo pallo;
-    private double korkeus;
-    private double leveys;
+    private int korkeus;
+    private int leveys;
     
     public Kentta() {
         this.pallo = new Pallo();
-        this.korkeus = 30.0;
-        this.leveys = 60.0;
+        this.korkeus = 30;
+        this.leveys = 60;
     }
     
     public Pallo getPallo() {
         return this.pallo;
     }
 
-    public double kenellePiste() {
-        double kummalle = 0.0;
-        if (this.pallo.getX() < 2.0) {
-            kummalle = 1.0;
-        } else if (this.pallo.getX() > 58.0){
-            kummalle = 2.0;
+    public int kenellePiste() {
+        int kummalle = 0;
+        if (this.pallo.getX() < 2) {
+            kummalle = 1;
+        } else if (this.pallo.getX() > 58){
+            kummalle = 2;
         }
         return kummalle;
     }
-    public double getLeveys() {
+    public int getLeveys() {
         return this.leveys;
     }
     
-    public double getKorkeus() {
+    public int getKorkeus() {
         return this.korkeus;
     }
