+++ Pong-2014/src/main/java/pong2014/logiikka/Pelaaja.java
@@ -9,7 +9,7 @@ public class Pelaaja {
     private Maila maila;
     private int pisteet;
     
-    public Pelaaja(double paikka) {
+    public Pelaaja(int paikka) {
         this.maila = new Maila(paikka);
         this.pisteet = 0;
     }
@@ -33,12 +33,12 @@ public class Pelaaja {
     }
 
     void lyo(Pallo pallo) {
-        double korkeus = pallo.getY();
-        double mailanSijainti = this.maila.getPaikka();
-        if (korkeus >= mailanSijainti && korkeus <= mailanSijainti + 6.0) {
+        int korkeus = pallo.getY();
+        int mailanSijainti = this.maila.getPaikka();
+        if (korkeus >= mailanSijainti && korkeus <= mailanSijainti + 6) {
             pallo.kimpoaMailasta(1);
         } else {
-            pallo.setPaikka(-1.0);
+            pallo.setPaikka(-1);
         }
     }
